parsing configuration zones #26
not tested yet
