+++ src/control_unit/control_unit_A/danger_evaluation.h
@@ -1,0 +1,16 @@
+
+#include "forbidden_zones_config_parse.h"
+#include "logs.h"
+
+// Skontroluje, či bod patrí do jednej zakázanej zóny
+bool isPointInPolygon(AzimuthElevation* azimutElevation, const ForbiddenZone& polygon);
+
+// Skontroluje, či bod patrí do niektorej zakázanej zóny
+int checkForbiddenZone(AzimuthElevation* azimutElevation);
+
+int enteredForbidenZone(AzimuthElevation* azimutElevation){
+    //writeAlarmToLog(azimutElevation);
+    //startAlarm();
+    //disableMotors();
+    return 0;
+}
