+++ src/control_unit/control_unit_A/forbidden_zones_config_parse.h
@@ -1,9 +1,5 @@
 
-#ifndef FORBIDDEN_ZONES_H
-#define FORBIDDEN_ZONES_H
-
 #include "common_structs.h"
-#include "logs.h"
 #include <vector>
 
 // Zakázaná zóna (polygón)
@@ -14,17 +10,4 @@ using ForbiddenZones = std::vector<ForbiddenZone>;
 
 extern ForbiddenZones systemForbiddenZones;
 
-// Skontroluje, či bod patrí do jednej zakázanej zóny
-bool isPointInPolygon(AzimuthElevation* azimutElevation, const ForbiddenZone& polygon);
-
-// Skontroluje, či bod patrí do niektorej zakázanej zóny
-int checkForbiddenZone(AzimuthElevation* azimutElevation);
-
-int enteredForbidenZone(AzimuthElevation* azimutElevation){
-    writeAlarmToLog(azimutElevation);
-    startAlarm();
-    disableMotors();
-    return 0;
-}
-
-#endif // FORBIDDEN_ZONES_H
+int checkFileFormat(char* newConfiguration);
