+++ src/control_unit/control_unit_A/logs.h
@@ -10,7 +10,6 @@ int writeHeaderIfNeeded(const char *filename, const char *header);
 int writeAEtoLog(AzimuthElevation *azimuthElevation);
 int writeChangeToLog(ChangeType changeType, const char *ip);
 int writeAlarmToLog(AzimuthElevation *azimuthElevation);
-int checkFileFormat(char* newConfiguration);
 int writeConfigAlarmAndIntervals(const char* data);
 int writeConfigZones(const char* zones);
 char* loadFile(const char* filePath);
